Restructured directory and component folders added. Program is tested and signup and login page still loads and runs.
