+++ src/components/WorkoutPage/Workout.jsx
@@ -1,4 +1,4 @@
-export default function Dash(){
+export default function Workout(){
     return(
         <div>
             <h1>Welcome Peasants</h1>
